# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenants/{tenantId}/events/ticket": {
      "post": {
        "description": "A browser `EventSource` cannot send an `Authorization` header, and the access token must never be put in a URL. Call this with the normal `Authorization` header, then open `GET …/events?ticket=<ticket>` within 30 seconds. The ticket works once; open a new one (this call again) for every reconnect. A server-side proxy that can set the header does not need tickets.",
        "operationId": "StaffEventsController_ticket",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamTicketResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A single-use ticket to open the event stream from a browser",
        "tags": [
          "events"
        ]
      }
    },
    "/v1/tenants/{tenantId}/events": {
      "get": {
        "description": "A `text/event-stream` for every staff role. Authentication: `Authorization: Bearer <token>` (from a server-side proxy) or `?ticket=` (direct browser use).\n\nEvents (the payload only has ids, so the UI refetches): `conversation.escalated`, `conversation.assigned`, `conversation.released`, `conversation.resolved` `{conversationId}`; `message.created` `{conversationId, messageId, authorType}`; `action.proposed` `{actionId, conversationId?}`; `notification.created` `{notificationId, type}` (only on the stream of the user it is for).\n\nControl events: `ready` (the stream is open), `resync` (your `Last-Event-ID` is too old or from before a restart: reload your lists), `closed` `{reason}` (`expired`, `revoked` when the user was disabled, `limit` when a sixth stream replaced this one). A `: ping` comment arrives every 25 seconds. Every data event has an `id`; send it back as `Last-Event-ID` to receive what you missed (the last 100 events are kept).\n\nAt most 5 streams per user; opening another closes the oldest. The fan-out lives in the memory of one backend instance: several instances need a shared bus (Redis, Phase 8). A stream ends after one hour; reconnect.",
        "operationId": "StaffEventsController_stream",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastEventId",
            "required": false,
            "in": "query",
            "description": "Resume point for clients that cannot send the `Last-Event-ID` header (a new `EventSource` after a reconnect).",
            "schema": {}
          },
          {
            "name": "ticket",
            "required": false,
            "in": "query",
            "description": "A single-use ticket from `POST …/events/ticket`, instead of the `Authorization` header. The access token itself is never accepted in the query string.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "retry: 3000\n\nevent: ready\ndata: {\"heartbeatSeconds\":25}\n\nid: 9f2a-1\nevent: conversation.escalated\ndata: {\"conversationId\":\"…\"}\n\n"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Live events for the dashboard (server-sent events)",
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-tenant support platform: backend core",
    "description": "Tenants, staff authentication, invitations and roles, audit log, translations, end customers. Errors always have the shape `{ statusCode, code, message }`; lists are `{ data, total, skip, take }`.",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StreamTicketResponse": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "description": "Single use, valid for 30 seconds: open the stream with `?ticket=<this>` right away."
          },
          "expiresInSeconds": {
            "type": "number",
            "example": 30
          }
        },
        "required": [
          "ticket",
          "expiresInSeconds"
        ]
      },
...
    }
  }
}